# packages/openclaw/contracts/relay-openapi.yaml
openapi: 3.1.0
info:
  title: Relay API
  version: 1.0.0
  description: |-
    Create conversations between users and agents.

    Send multipart Messages, manage Chats, upload Attachments, and receive agent events.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.relayapp.im
    description: Relay API
security:
  - BearerAuth: []
...
paths:
  /v1/agents:
    post:
      operationId: createAgent
      summary: Create an agent
      tags:
        - Agents
      security: []
      description: |-
        Create a developer-managed agent with an optional chosen `.dev` Handle,
        display name, and image. Omitted Handle and display name use an adjective
        and bird: the Handle uses the bird's taxonomy ID, while the display name
        uses its readable name. A color can resolve a generated-name collision.
        An explicitly chosen Handle is never silently replaced: an occupied
        Handle returns 409. Omitted image uses a handle-deterministic monogram recipe and its rendered PNG.

        Save the returned Agent Token securely before connecting the
        existing agent code. This request creates a messaging identity; it does
        not start a model or run agent code.

        This bootstrap request does not require a bearer token. It is limited
        to 10 creations per caller address per day. On 429, respect Retry-After.
        Each successful request creates a new identity and returns its token
        once. Do not automatically retry an uncertain creation response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
            example:
              token_name: Relay CLI
      responses:
        "201":
          description: Agent created; save the returned Agent Token.
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAgentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: The explicitly chosen Handle is already in use.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request body exceeds 8192 bytes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Agent creation limit reached.
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema:
                type: integer
                minimum: 1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          description: An available Agent Handle could not be allocated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/chats/{chatId}/read:
    post:
      operationId: markChatAsRead
      summary: Mark chat as read
      description: Explicitly mark visible Messages in a Chat as Read. Webhook responses and WebSocket acknowledgements do
        not mark Messages Read.
      tags:
        - Chats
      security:
        - BearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          description: Unique identifier of the chat
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Chat marked as read successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/chats/{chatId}/messages:
    post:
      operationId: sendMessageToChat
      summary: Send a message to an existing chat
      tags:
        - Messages
      security:
        - BearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          description: Unique identifier of the chat
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: false
          description: Optional 1–255 character key for safe retries. If message.idempotency_key is also present, the two values
            must match.
          schema:
            type: string
            minLength: 1
            maxLength: 255
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageToChatRequest"
      responses:
        "202":
          description: Message accepted for delivery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: |-
            The send is not permitted. A Chat whose member set has no agent in
            it is rejected with error code `2028`. Who may reach whom was
            decided when the Chat was created; every active member may write
            to it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                chat_without_agent:
                  summary: A chat must include an agent
                  value:
                    error:
                      status: 403
                      code: 2028
                      message: A chat must include an agent.
                      doc_url: https://docs.relayapp.im/error/codes/2xxx/2028
                    success: false
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Chat is unavailable, a mention was used in a non-group chat, or a mention target is no longer in the chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
      description: |-
        Send a Message to an existing Chat.

        Blocking is silent on this path. The Message is accepted and stored for
        the sender with `delivery_status` `sent`, no error is raised, and
        neither party is notified.

        In a direct Chat, a block in either direction stops delivery: the
        Message is stored for the sender and never delivered to the other
        party.

        In a group Chat, blocking is one way. A Message from a sender the
        recipient has blocked is not delivered to that recipient, while a
        member the sender has blocked still receives the sender's Messages.
        Every other member receives the Message as normal.

        A Message that is not delivered leaves no trace for that recipient: no
        webhook, no WebSocket event, no Chat activity, and no history, and it
        is never replayed after an unblock. The sender's `message.delivered`
        fires when at least one recipient was delivered.
    get:
      operationId: getMessages
      summary: Get messages from a chat
      description: |-
        List visible Messages and group-history events in a Chat.

        The default order is oldest first. Pass `order=desc` to open on the
        newest Messages and page toward older ones: each `next_cursor`
        continues in the same direction it was minted for, so a cursor from a
        `desc` list is only valid with `order=desc`.
      tags:
        - Messages
      security:
        - BearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          description: Unique identifier of the chat
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          description: Pagination cursor from previous next_cursor response
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of messages to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: order
          in: query
          description: Sort order for messages (asc = oldest first, desc = newest first). A cursor is only valid with the order it was returned for.
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
        - name: thread_originator_message_id
          in: query
          description: Return one thread only, in time order, with no group-history events. The result is the Message that
            opened the thread plus every Message whose `thread.originator_message_id` is that Message, so a reply to a reply is
            included. Answers 404 when the Message is not in this Chat or the caller cannot see it. A cursor from a thread
            list is only valid with the same thread.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetMessagesResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/webhook-subscriptions:
    post:
      operationId: createWebhookSubscription
      summary: Create a new webhook subscription
      tags:
        - Webhooks
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookSubscriptionRequest"
      responses:
        "201":
          description: Webhook subscription created successfully. The signing_secret is only returned in this response - store it
            securely.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscriptionCreatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AgentOnly"
        "500":
          $ref: "#/components/responses/InternalServerError"
      description: Create a signed webhook subscription.
    get:
      operationId: listWebhookSubscriptions
      summary: List all webhook subscriptions
      description: List webhook subscriptions for the authenticated agent.
      tags:
        - Webhooks
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of webhook subscriptions retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListWebhookSubscriptionsResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AgentOnly"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/contact_card:
    get:
      operationId: getContactCard
      summary: Get contact cards
      tags:
        - Contact Card
      security:
        - BearerAuth: []
      parameters:
        - name: handle
          in: query
          required: false
          description: Relay Handle. Agents can access only their own card.
          schema:
            type: string
      responses:
        "200":
          description: Contact card(s) returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetContactCardResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
      description: Retrieve the authenticated agent's Contact Card.
    post:
      operationId: setupContactCard
      summary: Setup contact card
      tags:
        - Contact Card
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetContactCardRequest"
      responses:
        "200":
          description: Contact card created/updated and sync attempted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetContactCardResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: The authenticated agent has no Contact card row to update (error code `2001`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
      description: Activate the authenticated agent's Contact Card.
    patch:
      operationId: updateContactCard
      summary: Update contact card
      tags:
        - Contact Card
      security:
        - BearerAuth: []
      parameters:
        - name: handle
          in: query
          required: true
          description: Relay Handle. Agents can access only their own card.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactCardRequest"
      responses:
        "200":
          description: Contact card updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetContactCardResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      description: Update the authenticated agent's Contact Card.
  /v1/websocket:
    get:
      operationId: connectAgentWebSocket
      summary: Connect an Agent WebSocket
      tags:
        - WebSocket
      security:
        - BearerAuth: []
      parameters:
        - name: observe
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: >-
            Read-only diagnostic event view. The ready frame includes
            observational:true. Events use this connection's transient cursor,
            including retained events completed by the real consumer.
            ACK and full_sync_complete frames are rejected. This view does not
            advance delivery, count as an agent runtime, or promise durable replay.
      responses:
        "101":
          description: WebSocket connection accepted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
      description: >-
        Upgrade wss://api.relayapp.im/v1/websocket with Authorization: Bearer
        <Agent Token>. If the Agent has any saved webhook subscription, Relay
        returns 409. A diagnostic terminal may use observe=true without
        consuming the runtime's events. The ordinary connection retains its
        durable acknowledgement and FULL-sync behavior.
components:
  schemas:
...
    CreateAgentRequest:
      type: object
      additionalProperties: false
      dependentRequired:
        image_recipe:
          - image_url
      properties:
        about:
          type: string
          minLength: 1
          maxLength: 60
          description: About text. Surrounding whitespace is trimmed.
        token_name:
          type: string
          minLength: 1
          maxLength: 80
          pattern: '^[^\u0000-\u001f\u007f]+$'
          default: Relay CLI
          description: A label for the issued token, not an agent name or owner identity.
        handle:
          type: string
          pattern: '^[a-z][a-z0-9_]{2,31}\.dev$'
          description: >-
            An available developer Handle, including .dev. Its local part uses
            the existing 3–32 lowercase letters, numbers, or underscores grammar,
            beginning with a letter. Omit for an assigned adjective/bird-ID Handle.
        first_name:
          type: string
          minLength: 1
          maxLength: 30
          pattern: '^[^\u0000-\u001f\u007f]+$'
          description: Display name. Surrounding whitespace is trimmed. Omit for an assigned adjective/bird name.
        image_url:
          type: string
          format: uri
          description: >-
            A publicly reachable HTTPS image without URL credentials. Relay uses
            its existing image ingestion and permanent-image storage. Omit for
            a handle-deterministic monogram recipe and its rendered PNG.
        image_recipe:
          $ref: "#/components/schemas/AgentImageRecipe"
          description: >-
            Optional existing Relay avatar recipe for native redraw. Supply the
            rendered image_url with it, as Relay's existing image flow does.
            A recipe alone is not an image-rendering service.
    AgentImageRecipe:
      oneOf:
        - $ref: "#/components/schemas/AgentMonogramImageRecipe"
        - $ref: "#/components/schemas/AgentEmojiImageRecipe"
        - $ref: "#/components/schemas/AgentPhotoImageRecipe"
...
    CreateAgentResponse:
      type: object
      additionalProperties: false
      required:
        - agent
        - secret
        - share_url
      properties:
        agent:
          allOf:
            - $ref: "#/components/schemas/ContactCardItem"
            - type: object
              properties:
                kind:
                  const: agent
        secret:
          type: string
          pattern: '^rel_token_[A-Za-z0-9]{43}$'
          description: The Agent Token, returned only in this creation response.
          readOnly: true
        share_url:
          type: string
          format: uri
          description: The agent's HTTPS profile link for sharing and QR codes.
...
    CreateWebhookSubscriptionRequest:
      type: object
      required:
        - target_url
        - subscribed_events
      properties:
        target_url:
          type: string
          format: uri
          description: URL where webhook events will be sent. Must be HTTPS.
        subscribed_events:
          type: array
          description: List of event types to subscribe to
          minItems: 1
          items:
            $ref: "#/components/schemas/WebhookEventType"
...
    ErrorResponse:
      type: object
      required:
        - error
        - success
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
        success:
          type: boolean
          description: Always false for error responses
        trace_id:
          type: string
          description: Unique trace ID for request tracing and debugging
    GetContactCardResponse:
      type: object
      required:
        - contact_cards
      properties:
        contact_cards:
          type: array
          items:
            $ref: "#/components/schemas/ContactCardItem"
    GetMessagesResult:
      type: object
      required:
        - messages
      properties:
        messages:
          type: array
          description: List of messages
          items:
            $ref: "#/components/schemas/Message"
        next_cursor:
          type:
            - string
            - "null"
          description: |
            Cursor for fetching the next page of results.
            Null if there are no more results to fetch.
            Pass this value as the `cursor` parameter in the next request.
...
    ListWebhookSubscriptionsResult:
      type: object
      required:
        - subscriptions
      properties:
        subscriptions:
          type: array
          description: List of webhook subscriptions
          items:
            $ref: "#/components/schemas/WebhookSubscriptionResponse"
...
    SendMessageResponse:
      type: object
      description: Response for sending a message to a chat
      required:
        - chat_id
        - message
      properties:
        chat_id:
          type: string
          format: uuid
          description: Unique identifier of the chat this message was sent to
        message:
          $ref: "#/components/schemas/SentMessage"
...
    SendMessageToChatRequest:
      type: object
      required:
        - message
      properties:
        message:
          $ref: "#/components/schemas/MessageContent"
...
    SetContactCardResponse:
      type: object
      required:
        - handle
        - first_name
        - is_active
        - kind
        - last_name
        - image_url
      properties:
        first_name:
          type: string
          description: First name on the contact card
        last_name:
          type:
            - string
            - "null"
        image_url:
          type:
            - string
            - "null"
          format: uri
        is_active:
          type: boolean
          description: Whether the contact card was successfully applied to the device
        handle:
          type: string
          description: Relay Handle associated with this contact card.
        kind:
          type: string
          enum:
            - user
            - agent
          description: Whether the Contact is a user or an agent.
...
    WebhookSubscriptionCreatedResponse:
      type: object
      description: Response returned when creating a webhook subscription. Includes the signing secret which is only shown once.
      required:
        - id
        - target_url
        - signing_secret
        - subscribed_events
        - is_active
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique identifier for the webhook subscription
        target_url:
          type: string
          format: uri
          description: URL where webhook events will be sent
        signing_secret:
          type: string
          description: Secret for verifying webhook signatures. Store this securely - it cannot be retrieved again.
        subscribed_events:
          type: array
          description: List of event types this subscription receives
          items:
            $ref: "#/components/schemas/WebhookEventType"
        is_active:
          type: boolean
          description: Whether this subscription is currently active
        created_at:
          type: string
          format: date-time
          description: When the subscription was created
        updated_at:
          type: string
          format: date-time
          description: When the subscription was last updated
...
  responses:
    AgentOnly:
      description: Forbidden - the route requires an active agent principal (error code `2003`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request - validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: The request conflicts with the current state of the resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden - no access to this resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Unauthorized:
      description: Unauthorized - missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token authentication. Include your API token in the Authorization header.

        Format: `Authorization: Bearer <your-token>`
...